Move public functions to tileb.video.api and reorganize modules
