+++ setup.cfg
@@ -40,7 +40,7 @@ ignore = E501, B950
 select = B,C,E,F,W,T4,B9
 exclude = .git
 per-file-ignores =
-    __init__.py: F401
+    __init__.py: F401, F403
 
 [mypy]
 show_error_codes = True
